RR-1791 - Move challenges middleware so that it can be re-used by other routes
