+++ server/routes/profile/challenges/index.ts
@@ -2,7 +2,7 @@ import { Router } from 'express'
 import asyncMiddleware from '../../../middleware/asyncMiddleware'
 import ChallengesController from './challengesController'
 import { Services } from '../../../services'
-import retrieveCurrentChallenges from '../middleware/retrieveCurrentChallenges'
+import retrieveChallenges from '../../middleware/retrieveChallenges'
 import retrieveAlnScreeners from '../../middleware/retrieveAlnScreeners'
 import retrievePrisonsLookup from '../../middleware/retrievePrisonsLookup'
 import retrieveEducationSupportPlanLifecycleStatus from '../middleware/retrieveEducationSupportPlanLifecycleStatus'
@@ -15,7 +15,7 @@ const challengesRoutes = (services: Services): Router => {
     .get('/', [
       retrieveEducationSupportPlanLifecycleStatus(educationSupportPlanService),
       retrievePrisonsLookup(prisonService),
-      retrieveCurrentChallenges(challengeService),
+      retrieveChallenges(challengeService),
       retrieveAlnScreeners(additionalLearningNeedsService),
       asyncMiddleware(controller.getChallengesView),
     ])
